resource/acl: add a few missing options (#9)
* dep: update dependencies

* Use the provider scaffolding setup

* resource/acl: add a few missing options

* resource/acl: implement logic to keep rules ordered

To make this resource work correctly with how the Cisco ASA uses the position field (as an `insert at` value), we needed to switch from using a `set` to using a `list`.

Next to that we needed to implement logic that understands how this positioning works inside an Cisco ASA and can set the position fields to the correct `insert at` position.

By adding this logic, we prevent the need to recreate all rules when a single rule is (for example) inserted at position 1. Instead the resource will now understand that when it inserts the new rule all the other rules are pushed up 1 position and so those rules do not have to be recreated in order to get to their desired postitions.

This also means that the position field is no longer available as a config value. Instead the order used to define the rules in your config, will define the order of the rules in the Cisco ASA.